Fix #750: [Sim] Refactor plots in sim_engine
