+++ pdr_backend/sim/test/test_sim_plot_state.py
@@ -1,0 +1,1 @@
+# do nothing here, it's all tested in test_sim_engine.py
